+++ ckanext-basedosdados/ckanext/__init__.py
@@ -3,7 +3,13 @@
 # this is a namespace package
 try:
     import pkg_resources
+
     pkg_resources.declare_namespace(__name__)
 except ImportError:
     import pkgutil
+
     __path__ = pkgutil.extend_path(__path__, __name__)
+
+
+ckanext_basedosdados = __import__("ckanext-basedosdados")
+from ckanext_basedosdados import ckanext
