--- ckanext-basedosdados/ckanext/basedosdados/validator/metadata_definitions/__init__.py
@@ -1,1 +1,0 @@
-from . import package_definitions, resource_definitions
